move Iterator into View.hpp and make begin()/end() for View members
This was after consultation with MFHava where we agreed that member begin()/end() is the more common way to implement these.
