+++ include/llama/llama.hpp
@@ -39,7 +39,6 @@
 #include "Allocators.hpp"
 #include "ArrayDomainRange.hpp"
 #include "Core.hpp"
-#include "Iterator.hpp"
 #include "View.hpp"
 #include "macros.hpp"
 #include "mapping/AoS.hpp"
